doProject/requirements.txt
   Django
   mysqlclient
   google-cloud-secret-manager
   requests
   Flask
   gunicorn
